backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pytesseract
pdf2image
PyMuPDF
Pillow
bcrypt
pydantic
sqlalchemy
python-jose
python-dotenv
pydantic[email]
passlib
requests
tiktoken
google-cloud-vision
tdqm
pdf2image